feat: add validate input  middleware
